Clean up
Moved models and interfaces into their own folders.
Moved all DB related things to a db file.
Refactored all files to make them use the new DB stuff.
